+++ docs/releasenotes.md
@@ -4,7 +4,7 @@
 * Savegames for *DOOM II: No Rest For The Living* are now saved in the correct folder, rather than in the same folder as *DOOM II: Hell On Earth* savegames.
 * A bug has been fixed whereby the frames per second displayed by enabling the `vid_showfps` CVAR wouldn’t display correctly in some instances.
 * The fuzz effect of spectres no longer bleeds into the top of the status bar, or into a higher sector when partially obscured by it.
-* The `+mouselook` action may now be bound to a key or a mouse button. Mouselook will then only work while that control is held down (or if the `mouselook` CVAR is `on`).
+* The `+mouselook` action may now be bound to a control. Mouselook will then only work while that control is held down (or if the `mouselook` CVAR is `on`).
 
 ---